Sample client done
